# swagger.yaml
swagger: '2.0'
info: 
  version: 0.0.1
  title: GAAP API
host: 'localhost:5000'
basePath: /api
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  JWT: 
    type: apiKey
    in: header
    name: auth
security:
  - JWT: []
parameters: 
  pageParam:
    name: page
    in: query
    description: The page number
    required: false
    type: integer
  pageSizeParam:
    name: pageSize
    in: query
    description: Max numer of elements per page
    required: false
    type: integer
  nameParam:
    name: name
    in: query
    description: name of the collaborator
    type: string
paths:
...
  /department:
    x-swagger-router-controller: DepartmentController
    x-typeorm-repository-name: Department
    get:
      x-security-privilege: 'department:read'
      operationId: consultDepartment
      description: consult department list
      tags:
        - department
      responses:
        '200':
          description: successful
          schema:
            $ref: '#/definitions/DepartmentResponse'
        '404':
          description: Test unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Test unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
  /department/{id}:
    x-swagger-router-controller: DepartmentController
    x-typeorm-repository-name: Department   
    get:
      x-security-privilege: 'department:read'
      operationId: cunsultStaffByDepartment
      description: consult staff by department
      tags:
        - department
      parameters:
        - name: id
          type: string
          format: uuid
          in: path
          required: true
      responses:
        '200':
          description: successful consult
          schema:
            $ref: '#/definitions/StaffResponse'
        '404':
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
  #---------------------//Department----------------
  #---------------------Collaborator----------------
...
  /section:
    x-swagger-router-controller: SectionController
    x-typeorm-repository-name: Section   
    get:
      x-security-privilege: 'section:read'
      operationId: consultSections
      description: consult sections (3 sections total) with number of collaborators and pagination based on page size
      tags:
        - section
      responses:
        '200':
          description: successful consult
          schema:
            type: object
            properties:
              sectionId:
                type: string
                format: uuid
              sectionName:
                type: string
        '404':
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'  
  /section/{id}:
    x-swagger-router-controller: SectionController
    x-typeorm-repository-name: collaborator   
    get:
      x-security-privilege: 'section:read'
      operationId: consultSectionPage
      description: return the page of collaborators of selected section (3 sections)
      tags:
        - section
      parameters:
        - name: id
          description: id of the section
          type: string
          format: uuid
          in: path
          required: true
        - $ref: '#/parameters/pageSizeParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/nameParam'
      responses:
        '200':
          description: successful consult
          schema:
            type: object
            properties:
              pageSize:
                type: number
              sectionCount: 
                type: number
              pageCount:
                type: number 
        '404':
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
  /sectionName/{id}:
    x-swagger-router-controller: SectionController
    x-typeorm-repository-name: collaborator     
    get:
      x-security-privilege: 'section:read'
      operationId: consultSectionPageName
      description: return the page of collaborators of selected section (3 sections)
      tags:
        - section
      parameters:
        - name: id
          description: id of the section
          type: string
          format: uuid
          in: path
          required: true
        - $ref: '#/parameters/nameParam'
      responses:
        '200':
          description: successful consult
          schema:
            type: object
            properties:
              pageSize:
                type: number
              sectionCount: 
                type: number
              pageCount:
                type: number 
        '404':
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
  /sectioncount/{id}:
    x-swagger-router-controller: SectionController
    x-typeorm-repository-name: Section   
    get:
      x-security-privilege: 'collaborator:read'
      operationId: consultSectionPagination
      description: consult section with number of collaborators on that section and number of pages needed based on pagesize
      tags:
        - section
      parameters:
        - name: id
          description: Id of the section to consult
          type: string
          format: uuid
          in: path
          required: true
        - $ref: '#/parameters/pageSizeParam'
      responses:
        '200':
          description: successful consult
          schema:
            type: object
            properties:
              sectionId:
                type: string
                format: uuid
              sectionCount:
                type: number
              sectionPages:
                type: number
        '404':
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description:  unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
#---------------//Section--------------------
#---------------Callendar--------------------
  /calendar:
    x-swagger-router-controller: CalendarController
    x-typeorm-repository-name: Calendar
    post:
      x-security-privilege: 'calendar:write'
      operationId: createCalendar
      description: create new event in calendar
      tags:
        - calendar
      parameters:
        - name: calendar
          in: body
          required: true
          description: the event to create
          schema:
            type: object
            properties:
              date:
                type: string
                format: date
              srcimg:
                type: string
              description:
                type: string 
              address:
                type: string
              title:
                type: string
      responses:
        '201':
          description: successful
          schema:
            $ref: '#/definitions/CalendarResponse'
        '404':
          description: unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'  
    get:
      x-security-privilege: 'calendar:read'
      operationId: consultCalendar
      description: consult calendar with pagination
      tags:
        - calendar
      parameters:
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/pageSizeParam'
      responses:
        '200':
          description: successful
          schema:
            $ref: '#/definitions/CalendarResponse'
        '404':
          description: unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'            
  /calendar/pages:
    x-swagger-router-controller: CalendarController
    x-typeorm-repository-name: Calendar
    get:
      x-security-privilege: 'calendar:read'
      operationId: consultCalendarPages
      description: consult calendar with pagination
      tags:
        - calendar
      parameters:
        - $ref: '#/parameters/pageSizeParam'
      responses:
        '200':
          description: successful
          schema:
            $ref: '#/definitions/CalendarResponse'
        '404':
          description: unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: unsuccessful
          schema:
            $ref: '#/definitions/ErrorResponse'            
#---------------//Calendar--------------------
#---------------definitions------------------
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  DepartmentResponse:  
    type: object
    properties:
        departmentId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string 
...
  CalendarResponse:  
    type: object
    properties:
        calendarId:
          type: string
          format: uuid
        date:
          type: string
        srcimg:
           type: string
        description:
           type: string 
        address:
           type: string
        title:
           type: string
  StaffResponse:
    type: object
    properties:
        staffId:
          type: string
          format: uuid
        email:
          type: string
        name:
           type: string